refactor itoperator into smaller classes
